Moved UILessCodeToTokenExchanger to "RunOnLoad". Also added "enter" keyboard shortcut to BrowserAuthDialog
